Create README.md
UsePython(thematplotlib.pyplotmodule) to plot the curves of sinxand cosxforx∈[−2π,2π] on the same graph.  You need to specify the title of the graph as “sinxvs cosxon [−2π,2π]”, x-label as “value of x”, y-label as “value of y”; The curve of sinxshould bein red solid line with circles marked at sampled values, the curve of cosxshould be in bluedashed line with square marked at sample values.  0.1 is the suggested distance between twosample points.  Print and hand in the graph and thePythoncode.

 Prepare a text file named “input.txt” with two columns.  The first column is 1,2,3,···,9and the second column is 1.1,2.2,3.3,···,9.9.  UsePythonto read data from “input.txt” andwrite calculated data to a new file “output.txt”.  The data in “output.txt” are composed offour columns: the first two columns are the same as those from “input.txt”, the third columnis the square root of the first column with 8 digits after the decimal point (e.g.  1.41421356),and the last column is the 10th power of the second column in scientific notation (exponential)with 4 digits after the decimal point (9.0438e+09).  Print and hand in the output.txt fileand thePythoncode.